sensegate-device/requirements.txt
Flask==3.1.0
PyYAML==6.0.2
requests==2.32.3
numpy==2.2.4
opencv-python==4.11.0.86
waitress==3.0.2
